DS-1064: Extract generic parts out of "AwsClients" into AwsClientFactoryBase
... to be re-usable by other client implementations and to prevent common mistakes
