+++ xyz-jobs/xyz-job-service/src/main/java/com/here/xyz/jobs/steps/execution/CleanUpExecutor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2024 HERE Europe B.V.
+ * Copyright (C) 2017-2025 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@
  */
 package com.here.xyz.jobs.steps.execution;
 
-import static com.here.xyz.jobs.util.AwsClients.cloudwatchEventsClient;
+import static com.here.xyz.jobs.util.AwsClientFactory.cloudwatchEventsClient;
 import static java.util.concurrent.TimeUnit.MINUTES;
 
 import com.here.xyz.jobs.Job;
