+++ xyz-jobs/xyz-job-service/src/main/java/com/here/xyz/jobs/steps/execution/StateMachineExecutor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2024 HERE Europe B.V.
+ * Copyright (C) 2017-2025 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@
 
 package com.here.xyz.jobs.steps.execution;
 
-import static com.here.xyz.jobs.util.AwsClients.sfnClient;
+import static com.here.xyz.jobs.util.AwsClientFactory.sfnClient;
 import static software.amazon.awssdk.services.sfn.model.StateMachineType.EXPRESS;
 import static software.amazon.awssdk.services.sfn.model.StateMachineType.STANDARD;
