+++ xyz-jobs/xyz-job-steps/src/main/java/com/here/xyz/jobs/util/AwsClientFactory.java
@@ -19,56 +19,39 @@
 
 package com.here.xyz.jobs.util;
 
-import com.here.xyz.jobs.steps.Config;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.awscore.client.builder.AwsClientBuilder;
-import software.amazon.awssdk.regions.Region;
+import com.here.xyz.util.service.aws.AwsClientFactoryBase;
 import software.amazon.awssdk.services.cloudwatchevents.CloudWatchEventsClient;
 import software.amazon.awssdk.services.emrserverless.EmrServerlessClient;
 import software.amazon.awssdk.services.sfn.SfnAsyncClient;
 import software.amazon.awssdk.services.sfn.SfnClient;
 
-public class AwsClients {
+public class AwsClientFactory extends AwsClientFactoryBase {
   private static SfnClient sfnClient;
   private static SfnAsyncClient asyncSfnClient;
   private static CloudWatchEventsClient cloudwatchEventsClient;
   private static EmrServerlessClient emrServerlessClient;
 
-  private static <T extends AwsClientBuilder> T prepareClientForLocalStack(T builder) {
-    if (isLocal()) {
-      builder.endpointOverride(Config.instance.LOCALSTACK_ENDPOINT);
-      builder.credentialsProvider(StaticCredentialsProvider.create(AwsBasicCredentials.create("localstack","localstack")));
-      builder.region(Region.of(Config.instance.AWS_REGION));
-    }
-    return builder;
-  }
-
   public static SfnClient sfnClient() {
     if (sfnClient == null)
-      sfnClient = prepareClientForLocalStack(SfnClient.builder()).build();
+      sfnClient = prepareClient(SfnClient.builder()).build();
     return sfnClient;
   }
 
   public static SfnAsyncClient asyncSfnClient() {
     if (asyncSfnClient == null)
-      asyncSfnClient = prepareClientForLocalStack(SfnAsyncClient.builder()).build();
+      asyncSfnClient = prepareClient(SfnAsyncClient.builder()).build();
     return asyncSfnClient;
   }
 
   public static CloudWatchEventsClient cloudwatchEventsClient() {
     if (cloudwatchEventsClient == null)
-      cloudwatchEventsClient = prepareClientForLocalStack(CloudWatchEventsClient.builder()).build();
+      cloudwatchEventsClient = prepareClient(CloudWatchEventsClient.builder()).build();
     return cloudwatchEventsClient;
   }
 
   public static EmrServerlessClient emrServerlessClient() {
     if (emrServerlessClient == null)
-      emrServerlessClient = prepareClientForLocalStack(EmrServerlessClient.builder()).build();
+      emrServerlessClient = prepareClient(EmrServerlessClient.builder()).build();
     return emrServerlessClient;
-  }
-
-  private static boolean isLocal() {
-    return Config.instance.LOCALSTACK_ENDPOINT != null;
   }
 }
